refactor: extract toolbar callbacks into dedicated module
Move build_toolbar_keyboard, _send_toolbar_key, _handle_toolbar_*,
and all provider row-2 definitions from screenshot_callbacks into
toolbar_callbacks.py — a single-responsibility module for /toolbar UI.
Update bot.py import, register module in callback_registry, update tests.
