+++ tests/ccgram/handlers/test_toolbar.py
@@ -1,6 +1,6 @@
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from ccgram.handlers.screenshot_callbacks import (
+from ccgram.handlers.toolbar_callbacks import (
     _send_toolbar_key,
     build_toolbar_keyboard,
 )
@@ -81,8 +81,8 @@ class TestBuildToolbarKeyboardRow2:
 
 
 class TestSendToolbarKey:
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=True)
-    @patch("ccgram.handlers.screenshot_callbacks.tmux_manager", new_callable=AsyncMock)
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=True)
+    @patch("ccgram.handlers.toolbar_callbacks.tmux_manager", new_callable=AsyncMock)
     async def test_mode_sends_shift_tab(self, mock_tmux, _mock_owns):
         query = AsyncMock()
         await _send_toolbar_key(
@@ -92,23 +92,23 @@ class TestSendToolbarKey:
         assert mock_tmux.send_keys.call_args[0][1] == "\x1b[Z"
         assert mock_tmux.send_keys.call_args[1]["literal"] is True
 
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=True)
-    @patch("ccgram.handlers.screenshot_callbacks.tmux_manager", new_callable=AsyncMock)
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=True)
+    @patch("ccgram.handlers.toolbar_callbacks.tmux_manager", new_callable=AsyncMock)
     async def test_regular_key_not_literal(self, mock_tmux, _mock_owns):
         query = AsyncMock()
         await _send_toolbar_key(query, 123, "tb:esc:@0", "tb:esc:", "Escape", "Esc")
         assert mock_tmux.send_keys.call_args[1]["literal"] is False
 
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=False)
-    @patch("ccgram.handlers.screenshot_callbacks.tmux_manager", new_callable=AsyncMock)
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=False)
+    @patch("ccgram.handlers.toolbar_callbacks.tmux_manager", new_callable=AsyncMock)
     async def test_not_owner_rejected(self, mock_tmux, _mock_owns):
         query = AsyncMock()
         await _send_toolbar_key(query, 123, "tb:esc:@0", "tb:esc:", "Escape", "Esc")
         mock_tmux.send_keys.assert_not_called()
         query.answer.assert_awaited_once_with("Not your session", show_alert=True)
 
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=True)
-    @patch("ccgram.handlers.screenshot_callbacks.tmux_manager", new_callable=AsyncMock)
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=True)
+    @patch("ccgram.handlers.toolbar_callbacks.tmux_manager", new_callable=AsyncMock)
     async def test_window_not_found(self, mock_tmux, _mock_owns):
         mock_tmux.find_window_by_id.return_value = None
         query = AsyncMock()
@@ -117,61 +117,58 @@ class TestSendToolbarKey:
 
 
 class TestHandleToolbarSend:
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=False)
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=False)
     async def test_not_owner_rejected(self, _mock_owns):
         query = AsyncMock()
         query.data = "tb:send:@0"
         update = MagicMock()
         context = MagicMock()
         context.user_data = {}
-        from ccgram.handlers.screenshot_callbacks import _handle_toolbar_send
+        from ccgram.handlers.toolbar_callbacks import _handle_toolbar_send
 
         await _handle_toolbar_send(query, 123, "tb:send:@0", update, context)
         query.answer.assert_awaited_once_with("Not your session", show_alert=True)
 
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=True)
-    @patch("ccgram.handlers.screenshot_callbacks.session_manager")
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=True)
+    @patch("ccgram.handlers.toolbar_callbacks.session_manager")
     async def test_no_cwd_shows_error(self, mock_sm, _mock_owns):
         mock_sm.get_window_state.return_value = None
         query = AsyncMock()
         update = MagicMock()
         context = MagicMock()
         context.user_data = {}
-        from ccgram.handlers.screenshot_callbacks import _handle_toolbar_send
+        from ccgram.handlers.toolbar_callbacks import _handle_toolbar_send
 
         await _handle_toolbar_send(query, 123, "tb:send:@0", update, context)
         query.answer.assert_awaited_once_with(
             "Working directory not available", show_alert=True
         )
 
-    @patch("ccgram.handlers.screenshot_callbacks.user_owns_window", return_value=True)
-    @patch("ccgram.handlers.screenshot_callbacks.thread_router")
-    @patch("ccgram.handlers.screenshot_callbacks.session_manager")
-    @patch("ccgram.handlers.screenshot_callbacks.safe_send", new_callable=AsyncMock)
-    @patch("ccgram.handlers.screenshot_callbacks.build_file_browser")
-    async def test_opens_file_browser(
-        self, mock_browser, mock_safe_send, mock_sm, mock_router, _mock_owns, tmp_path
-    ):
+    @patch("ccgram.handlers.toolbar_callbacks.user_owns_window", return_value=True)
+    @patch("ccgram.handlers.toolbar_callbacks.thread_router")
+    @patch("ccgram.handlers.toolbar_callbacks.session_manager")
+    async def test_opens_file_browser(self, mock_sm, mock_router, _mock_owns, tmp_path):
         ws = MagicMock()
         ws.cwd = str(tmp_path)
         mock_sm.get_window_state.return_value = ws
         mock_router.resolve_chat_id.return_value = 999
-        mock_browser.return_value = ("Browse files", MagicMock(), [tmp_path / "a.txt"])
 
         query = AsyncMock()
         update = MagicMock()
         update.effective_message.message_thread_id = 42
         context = MagicMock()
         context.user_data = {}
-        from ccgram.handlers.screenshot_callbacks import _handle_toolbar_send
+        from ccgram.handlers.toolbar_callbacks import _handle_toolbar_send
 
-        with patch(
-            "ccgram.handlers.screenshot_callbacks.get_thread_id", return_value=42
+        mock_open_browser = AsyncMock()
+        with (
+            patch("ccgram.handlers.toolbar_callbacks.get_thread_id", return_value=42),
+            patch("ccgram.handlers.send_command.open_file_browser", mock_open_browser),
         ):
             await _handle_toolbar_send(query, 123, "tb:send:@0", update, context)
 
-        mock_safe_send.assert_awaited_once()
-        call_kwargs = mock_safe_send.call_args
-        assert call_kwargs[0][1] == 999
-        assert call_kwargs[1]["message_thread_id"] == 42
+        mock_open_browser.assert_awaited_once()
+        call_args = mock_open_browser.call_args
+        assert call_args[0][1] == 999
+        assert call_args[0][2] == 42
         query.answer.assert_awaited_once_with()
